+++ CHANGELOG.md
@@ -11,8 +11,11 @@
 	- query monthly-cpu-stats (Thanks @gmauro!)
 	- total number of jobs with exit states for galaxxy instance (Thanks @bruggerk)
 	- query workflow-invocation-totals
+	- server subcommand exposing previously hidden functions
 - Removed:
 	- filter hexdecodelines, this is now replaced by a built-in postgres function
+- Changed:
+	- Refactored internal function search to unify local, query, mutate, with the rest.
 
 # 17